Fix tampilan pembayaran ketika tambah portofolio, fix id portofolio, tambah change listener portofolio, tambah format rupiah, ganti tipe data harga ke long

+++ app/src/main/java/com/example/ternavest/model/Laporan.kt
@@ -10,7 +10,7 @@ data class Laporan(
         var judulLaporan : String? = "",
         var deskripsiLaporan : String? = "",
         var tanggal : String? = "",
-        var pemasukan  : Int? = 0,
-        var pengeluaran : Int? = 0,
+        var pemasukan  : Long? = 0,
+        var pengeluaran : Long? = 0,
         var photoLaporan : String? = ""
 ) : Parcelable
